+++ ii/Program.cs
@@ -8,6 +8,7 @@ using FellowOakDicom;
 using IsIdentifiable.Options;
 using IsIdentifiable.Redacting;
 using IsIdentifiable.Runners;
+using NLog;
 using System.Linq;
 using YamlDotNet.Serialization;
 
@@ -133,7 +134,10 @@ public static class Program
 
     private static int Run(IsIdentifiableDicomFileOptions opts)
     {
-        Inherit(opts);
+        var result = Inherit(opts);
+
+        if (result != 0)
+            return result;
 
         using var runner = new DicomFileRunner(opts)
         {
@@ -150,7 +154,10 @@ public static class Program
         ImplementationManager.Load<PostgreSqlImplementation>();
         ImplementationManager.Load<OracleImplementation>();
 
-        Inherit(opts);
+        var result = Inherit(opts);
+
+        if (result != 0)
+            return result;
 
         using var runner = new DatabaseRunner(opts)
         {
@@ -160,7 +167,10 @@ public static class Program
     }
     private static int Run(IsIdentifiableFileOptions opts)
     {
-        Inherit(opts);
+        var result = Inherit(opts);
+
+        if (result != 0)
+            return result;
 
         using var runner = new FileRunner(opts)
         {
@@ -185,12 +195,14 @@ public static class Program
             opts.InheritValuesFrom(GlobalOptions.IsIdentifiableReviewerOptions);
         }
     }
-    private static void Inherit(IsIdentifiableBaseOptions opts)
+    private static int Inherit(IsIdentifiableBaseOptions opts)
     {
         if (GlobalOptions?.IsIdentifiableOptions != null)
         {
             opts.InheritValuesFrom(GlobalOptions.IsIdentifiableOptions);
         }
+
+        return opts.UpdateConnectionStringsToUseTargets(out _);
     }
 
 }